+++ src/py_anim_bindings/py_bindings.h
@@ -7,3 +7,4 @@
 #include "py_handle_mode.h"
 #include "py_function.h"
 #include "py_extend.h"
+#include "py_range_end.h"